Keyboard Manager and Price Update
A keyboard manager script was added. This is used to retrieve and store the data from the Keyboard Scriptable object and is accessed by the UI scripts to display each component in the UI.

Price update in price update script. Currently using the Update method but will implement Mediator and Observer pattern.

Added Scripts:
- KeyboardManager.cs
- EventManager.cs

Changed Scripts:
- KeycapsUI.cs
- SwitchesUI.cs
- ModelColourUI.cs
- KeycapsUIController.cs
- SwitchesUIController.cs
- ModelColourUI.cs
- PriceUpdateUI.cs
- Keyboard.cs
- KeyboardSelection.cs